Proof of concept network rendering
Where to go from here?
Open up GLHCK API to public so that the networking
cane be done in seperate library.

Optimize the library so that changes are only being sent.
Maybe initial geometry data compression?

The renderer isn't supposed to be truly realtime or fast,
only reliable to synchorize assets beetwen 3D tools and renderer.